# Conflicts:
#	ProjectNT/Assets/03.Code/Scripts/Effects/EffectManager.cs
#	ProjectNT/Assets/04.Level/Prefabs/GameScene/MapPivot.prefab
#	ProjectNT/Assets/04.Level/Scenes/LobbyScene.unity
#	ProjectNT/ProjectSettings/EditorBuildSettings.asset

+++ ProjectNT/Assets/03.Code/Scripts/_Editor/Public/BeatMapData.cs
@@ -33,8 +33,8 @@ public class UpperNodeData
 public class SongData
 {
     public float songLength;
-    public int phase2;
-    public int phase3;
+    public float phase2;
+    public float phase3;
 }
 
 //그리드의 정보
